backend/src/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
langchain-core==0.3.0
langchain-openai==0.2.0
langgraph==0.2.22
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
gunicorn==21.2.0
google-cloud-logging==3.9.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0